+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Unreleased
 ----------
-- Added support for newtype structs and newtype enum variants to
-  `to_variant_name`
+- Added support for newtype structs, newtype enum variants, and tuple
+  enum variants to `to_variant_name`
 
 
 0.1.0